requirements.txt
Django==5.2.11
djangorestframework==3.16.1
psycopg2-binary==2.9.11
gunicorn==25.1.0
requests==2.32.5